polymarket_arbitrage/requirements.txt
py-clob-client>=0.34.5
requests>=2.31.0
websockets>=12.0
python-dotenv>=1.0.0
